There're a lot of changes done and tweaking made to this commit that I can't explain anymore Again, I've been very burnt out lately this week. I can't do much and only happen to commit myself to the worst, and miss school... (I only went this tuesday) I'm still glad though that I've been able to work around a freezing bug involving sre_defer_response (now sre_defer_res) and window resizing on WIndows. More update info again in CHANGELOG.txt Some changes may look weird or stupid, that's because I tried compiling on MinGW, and it was annoying... I don't know if I ever want to support it
I'm planning on continuing my level editor soon so anything related to that (ui maybe or file manipulation) may change or be new

+++ src/base/fontatlas.cpp
@@ -113,7 +113,7 @@ void FontAtlas::render_text(const FontRenderData& renderdata, const FontRenderTe
             if (last_space)
             {
                 extentdiff = linedat->extent - lastextent_space;
-                spacediff = &renderdata.text[i] - last_space - 1;
+                spacediff = static_cast<int>(&renderdata.text[i] - last_space - 1);
                 linedat->num_chars -= spacediff;
                 linedat->extent = lastextent_space;
             }
